Multiple pages/modularity update
Made the program modular and able to support multiple pages. From now on
"npm install react-router-dom" will be mandatory to run the project.

Added a rudimentary "under construction" page for all pages that are not
yet completed.

Fixed nav menu secondary options being slightly misaligned from intended
position.

Previously existing functionality should be entirely unchanged.
