+++ frontend/src/Account.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Account(){
+	return(
+		<>
+			<p className="errorMessage">Account: Under construction, come back later!</p>
+		</>
+	);
+}
