+++ frontend/src/Calendar.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Calendar(){
+	return(
+		<>
+			<p className="errorMessage">Calendar: Under construction, come back later!</p>
+		</>
+	);
+}
