+++ frontend/src/Catalog.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Catalog(){
+	return(
+		<>
+			<p className="errorMessage">Catalog: Under construction, come back later!</p>
+		</>
+	);
+}
