+++ frontend/src/Clubs.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Clubs(){
+	return(
+		<>
+			<p className="errorMessage">Clubs: Under construction, come back later!</p>
+		</>
+	);
+}
