+++ frontend/src/Feedback.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Feedback(){
+	return(
+		<>
+			<p className="errorMessage">Feedback: Under construction, come back later!</p>
+		</>
+	);
+}
