+++ frontend/src/Footer.js
@@ -1,0 +1,12 @@
+import './App.css';
+import React from 'react';
+
+export default function Footer(){
+	return(
+		<>
+			<footer>
+				<p>© 2023 BClub</p>
+			</footer>
+		</>
+	);
+}
