+++ frontend/src/NoPage.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function NoPage(){
+	return(
+		<>
+			<p className="errorMessage">Error 404: Page not found. Please double check your URL!</p>
+		</>
+	);
+}
