+++ frontend/src/Profile.js
@@ -1,0 +1,10 @@
+import './App.css';
+import React from 'react';
+
+export default function Profile(){
+	return(
+		<>
+			<p className="errorMessage"> Profile: Under construction, come back later!</p>
+		</>
+	);
+}
